# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Blog Pessoal",
    "version": "1.0.0",
    "description": "API RESTful completa para um sistema de blog pessoal desenvolvida em Laravel com autenticação JWT como trabalho de conclusão da matéria Sistemas Distribuídos e API da PÓS em FullStack da Unimar.",
    "contact": {
      "name": "Fabiano Souza de Oliveira",
      "email": "fsouzaweb@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000/api",
      "description": "Servidor de Desenvolvimento"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Autenticação JWT. Use: Bearer {seu_token_jwt}"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {"type": "integer", "example": 1},
          "name": {"type": "string", "example": "Fabiano Souza"},
          "email": {"type": "string", "format": "email", "example": "fabiano@blog.com"},
          "created_at": {"type": "string", "format": "date-time"},
          "updated_at": {"type": "string", "format": "date-time"}
        }
      },
      "Category": {
        "type": "object",
        "properties": {
          "id": {"type": "integer", "example": 1},
          "name": {"type": "string", "example": "Laravel"},
          "slug": {"type": "string", "example": "laravel"},
          "created_at": {"type": "string", "format": "date-time"},
          "updated_at": {"type": "string", "format": "date-time"}
        }
      },
      "Post": {
        "type": "object",
        "properties": {
          "id": {"type": "integer", "example": 1},
          "title": {"type": "string", "example": "Meu Primeiro Post"},
          "content": {"type": "string", "example": "Conteúdo do post..."},
          "published_at": {"type": "string", "format": "date-time"},
          "author": {"$ref": "#/components/schemas/User"},
          "category": {"$ref": "#/components/schemas/Category"},
          "tags": {
            "type": "array",
            "items": {"$ref": "#/components/schemas/Tag"}
          },
          "created_at": {"type": "string", "format": "date-time"},
          "updated_at": {"type": "string", "format": "date-time"}
        }
      },
      "Tag": {
        "type": "object",
        "properties": {
          "id": {"type": "integer", "example": 1},
          "name": {"type": "string", "example": "Laravel"},
          "slug": {"type": "string", "example": "laravel"},
          "created_at": {"type": "string", "format": "date-time"},
          "updated_at": {"type": "string", "format": "date-time"}
        }
      },
...
    }
  },
...
  "paths": {
...
    "/posts": {
      "get": {
        "tags": ["Posts"],
        "summary": "Listar todos os posts",
        "description": "Retorna uma lista de posts com filtros opcionais",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "category_id",
            "in": "query",
            "description": "Filtrar por categoria",
            "schema": {"type": "integer"}
          },
          {
            "name": "user_id",
            "in": "query",
            "description": "Filtrar por autor",
            "schema": {"type": "integer"}
          },
          {
            "name": "published",
            "in": "query",
            "description": "Filtrar por status de publicação",
            "schema": {"type": "boolean"}
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de posts obtida com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {"$ref": "#/components/schemas/Post"}
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Posts"],
        "summary": "Criar novo post",
        "description": "Cria um novo post no blog",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "content", "category_id"],
                "properties": {
                  "title": {"type": "string", "example": "Meu Novo Post"},
                  "content": {"type": "string", "example": "Este é o conteúdo do meu post..."},
                  "category_id": {"type": "integer", "example": 1},
                  "published_at": {"type": "string", "format": "date-time", "example": "2024-01-15T10:00:00Z"},
                  "tag_ids": {
                    "type": "array",
                    "items": {"type": "integer"},
                    "example": [1, 2, 3]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Post criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {"$ref": "#/components/schemas/Post"}
                  }
                }
              }
            }
          },
          "422": {"description": "Dados de validação inválidos"}
        }
      }
    },
...
  }
}